Update the neuron type definition

+++ neuronet/neurons/neuron_linear.c
@@ -5,7 +5,7 @@ void neuron_linear_create(neuron_t *n, uint32_t num_inputs) {
     n->num_inputs = num_inputs;
     n->num_coeffs = num_inputs+1;
     n->inputs = (double*)calloc(n->num_inputs, sizeof(double));
-    n->input_indices = (double*)calloc(n->num_inputs, sizeof(uint32_t));
+    n->input_indices = (uint32_t*)calloc(n->num_inputs, sizeof(uint32_t));
     n->coeffs = (double*)calloc(n->num_coeffs, sizeof(double));
     n->coeffs_backup = (double*)calloc(n->num_coeffs, sizeof(double));
     n->coeffs_delta = (double*)calloc(n->num_coeffs, sizeof(double));
